add configuration for ssl certificate verification
This commits implements the support for configuring ssl certificate
verification. by default it keeps the current behaviour.

It only configures the :verify_ssl parameter for rest_client if
the instance of foreman_api explicitly configure it.

example to disable the ssl verification:

hosts = ForemanApi::Resources::Host.new(
  {
    :base_url   => 'https://foreman.example.com',
    :username   => 'user',
    :password   => 'password',
    :verify_ssl => false
  })

This is needed because when we use puppet ca, it is a self signed certificate,
and the api could not verify it.

thanks
Rosano